Initial virtual memory stub
This is the first commit of my recreational kernel project.
The entry point is init module with _start function. This module
sets up temporary memory mapping to quickly enable the kernel to run
at higher-half virtual address space. Then, the main function is
called with pointer to FDT and the initial page table. The page
module provides a primitive page table and entry structures. Some
functions are left unimplemented, but I commit them nonetheless to
remember what I should code next. The uart module is for debugging
only and will be rewritten in the near future.
Also note that the linker script was pieced together to "just work"
and is inefficient and maybe not even correct.

The next goal of the project is to implement a page-allocator and
remap the kernel in a more fine-grained manner.